+++ CHANGELOG.md
@@ -13,6 +13,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+## [1.123.0]
+
+### Added
+- Constructor-schemes: the ability to remove a scheme from the list of saved schemes
+- Constructor-schemes: token-action click:image
+- Constructor-schemes: import\export schemes
+- User's TTL settings
+
+### Fixed
+- constructor-schemes: import\export on click to "orderTo" button
+
 ## [1.122.0]
 
 ### Added